% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SkeletonCohortDiagnosticsStudy.R
\docType{package}
\name{SkeletonCohortDiagnosticsStudy-package}
\alias{SkeletonCohortDiagnosticsStudy}
\alias{SkeletonCohortDiagnosticsStudy-package}
\title{SkeletonCohortDiagnosticsStudy: A Study Package for Generating Cohort Diagnostics}
\description{
A package that generates cohort diagnostics for cohort definitions included in the package.
}
\keyword{internal}
